Fix platform-specific TravisCI badges
The TravisCI badges we have on the developer page are currently targeting the overall CI build status for Linux/macOS. The problem with this is that if one job is failing, they both will show failure status. This changes the badges to indicate the status for the specific platform build job. This was fixed on xqemu GitHub README a while back but never made it to the website.