Move udev rules comments to separate lines
Current build of `udev` in Debian Testing (243-9) doesn't handle comments the same way as before - they need a line of their own to be respected and not parsed as an argument.
As the comment texts are not considered valid key/value pairs `udev` fails to parse them and ignores the rules.
Even if this might be considered a bug in the Debian build the XQEMU docs wouldn't hurt from respecting this behavior.